Move the asm.ppc.gnu into the anal ##asm
