+++ libr/anal/meson.build
@@ -129,6 +129,10 @@ r_anal_sources = [
   join_paths('..','asm','arch','pic','pic_baseline.c'),
   join_paths('..','asm','arch','pic','pic_pic18.c'),
   join_paths('..','asm','arch','ppc','libvle','vle.c'),
+  join_paths('../asm/arch','ppc','gnu','ppc-dis.c'),
+  join_paths('../asm/arch','ppc','gnu','ppc-opc.c'),
+ # join_paths('arch','ppc','libvle','vle.c'),
+ # join_paths('arch','ppc','libps','libps.c'),
   join_paths('..','asm','arch','propeller','propeller_disas.c'),
   join_paths('..','asm','arch','rsp','rsp_idec.c'),
   join_paths('..','asm','arch','tms320','tms320_dasm.c'),
