+++ libr/anal/p/ppc_gnu.mk
@@ -1,4 +1,6 @@
 OBJ_PPC=anal_ppc_gnu.o
+OBJ_PPC+=../../asm/arch/ppc/gnu/ppc-dis.o
+OBJ_PPC+=../../asm/arch/ppc/gnu/ppc-opc.o
 
 STATIC_OBJ+=${OBJ_PPC}
 TARGET_PPC=anal_ppc_gnu.${EXT_SO}
