+++ libr/asm/meson.build
@@ -40,7 +40,6 @@ r_asm_sources = [
   join_paths('p','asm_or1k.c'),
   join_paths('p','asm_ppc_as.c'),
   join_paths('p','asm_ppc_cs.c'),
-  join_paths('p','asm_ppc_gnu.c'),
   join_paths('p','asm_propeller.c'),
   join_paths('p','asm_pyc.c'),
   join_paths('p','asm_riscv.c'),
@@ -66,6 +65,8 @@ r_asm_sources = [
   join_paths('arch','8051','8051_ass.c'),
   join_paths('arch','v850','gnu','v850-dis.c'),
   join_paths('arch','v850','gnu','v850-opc.c'),
+  join_paths('arch','ppc','libvle','vle.c'),
+  join_paths('arch','ppc','libps','libps.c'),
   #join_paths('arch','arc','gnu','arc-dis.c'),
   #join_paths('arch','arc','gnu','arc-ext.c'),
   #join_paths('arch','arc','gnu','arc-opc.c'),
@@ -110,10 +111,6 @@ r_asm_sources = [
   join_paths('arch','nios','gnu','nios2-dis.c'),
   join_paths('arch','nios','gnu','nios2-opc.c'),
   join_paths('arch','or1k','or1k_disas.c'),
-  join_paths('arch','ppc','gnu','ppc-dis.c'),
-  join_paths('arch','ppc','gnu','ppc-opc.c'),
-  join_paths('arch','ppc','libvle','vle.c'),
-  join_paths('arch','ppc','libps','libps.c'),
   join_paths('arch','propeller','propeller_disas.c'),
   join_paths('arch','pyc','opcode_all.c'),
   join_paths('arch','pyc','opcode_anal.c'),
